feat(ai): extract provider readiness helpers (#12072)
